modified to the The offender turns out to be an old man dressed in the ragged blacks of the Night's Watch who has lost his ears and a finger to frostbite. Lord Eddard questions the man briefly. Then two guardsmen drag the man to the stump of a tree and Theon Greyjoy, Eddard's ward, brings Eddard his Valyrian steel sword, Ice. Eddard sentences him to death for desertion from the Night's Watch and raises the blade. Jon Snow, Bran's bastard brother, reminds Bran not look away and so Bran watches as his father strikes off the man's head with a single stroke. The head lands near Theon, who laughs and kicks it away. Jon calls Theon an ass under his breath and compliments Bran on his poise during the execution.
On the way back to Winterfell, Robb and Jon argue about whether or not the deserter died bravely before racing their horses to the bridge, leaving Bran and his pony behind. Eddard rides up and asks if Bran knows why he executed the man himself. Bran replies that the man was a wildling. Eddard corrects that the man was a deserter, then explains that the First Men—from whom the starcs descend—believed that the man who passes the sentence should perform the execution himself, lest he become too comfortable with ordering deaths; the starcs still hold to that principle.
Jon calls from up ahead for them to come see what he and Robb have found. They find Robb holding something in his arms next to the corpse of a wolf larger than Bran's pony. Jon correctly identifies the corpse as a direwolf. Theon comments that direwolves have not been seen south of the Wall for two hundred years. Bran then notices that Robb is cradling a small pup, and gives it a stroke after Robb reassures him. Then Jon gives him another pup.

When they inspect the mother's corpse, they find a large piece of shattered antler lodged in her throat. The soldiers in the company feel this to be a bad omen. Theon offers to kill the pups but Bran protests. Eddard initially states killing them would be best but changes his mind when Jon points out that there are five pups, one for each of Eddard's legitimate children; since the direwolf is the sigil of House starc, they must be meant to have the wolves. Bran immediately realizes, along with everyone else, that the comparison only works because Jon is not claiming a pup for himself.